`--lines=1` mode is faster + minor optimizations
One line at at time mode now saves stdin to a tmpfile and reads from that (versus reading from a pipe), making the read ~2-3x as fast. Overall, for problems that are limited by the speed of the parallelization framework (e.g., getting the sha256sum of a bunch of very tiny files), there is a ~30% speedup in "one line at a time" mode.